+++ x/transferinject/types/expected_keepers.go
@@ -4,8 +4,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
 	chantypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
-	delayedacktypes "github.com/dymensionxyz/dymension/v3/x/delayedack/types"
-
 	rollapptypes "github.com/dymensionxyz/dymension/v3/x/rollapp/types"
 )
 
@@ -21,9 +19,9 @@ type RollappKeeper interface {
 		ctx sdk.Context,
 		packetData []byte,
 		raPortOnHub, raChanOnHub string,
-	) (data delayedacktypes.TransferData, err error)
+	) (data rollapptypes.TransferData, err error)
 	GetValidTransferFromSentPacket(
 		ctx sdk.Context,
 		packet chantypes.Packet,
-	) (data delayedacktypes.TransferData, err error)
+	) (data rollapptypes.TransferData, err error)
 }
